+++ h2/src/main/org/h2/engine/Database.java
@@ -42,9 +42,11 @@ import org.h2.message.DbException;
 import org.h2.message.Trace;
 import org.h2.message.TraceSystem;
 import org.h2.mvstore.MVStore;
+import org.h2.mvstore.db.LobStorageMap;
 import org.h2.mvstore.db.MVTableEngine;
 import org.h2.pagestore.PageStore;
 import org.h2.pagestore.WriterThread;
+import org.h2.pagestore.db.LobStorageBackend;
 import org.h2.result.LocalResultFactory;
 import org.h2.result.Row;
 import org.h2.result.RowFactory;
@@ -59,10 +61,8 @@ import org.h2.store.FileLock;
 import org.h2.store.FileLockMethod;
 import org.h2.store.FileStore;
 import org.h2.store.InDoubtTransaction;
-import org.h2.store.LobStorageBackend;
 import org.h2.store.LobStorageFrontend;
 import org.h2.store.LobStorageInterface;
-import org.h2.store.LobStorageMap;
 import org.h2.store.fs.FileUtils;
 import org.h2.table.Column;
 import org.h2.table.IndexColumn;
