+++ h2/src/main/org/h2/tools/Recover.java
@@ -42,6 +42,7 @@ import org.h2.mvstore.MVMap;
 import org.h2.mvstore.MVStore;
 import org.h2.mvstore.MVStoreTool;
 import org.h2.mvstore.StreamStore;
+import org.h2.mvstore.db.LobStorageMap;
 import org.h2.mvstore.db.ValueDataType;
 import org.h2.mvstore.tx.TransactionMap;
 import org.h2.mvstore.tx.TransactionStore;
@@ -49,6 +50,7 @@ import org.h2.pagestore.Page;
 import org.h2.pagestore.PageFreeList;
 import org.h2.pagestore.PageLog;
 import org.h2.pagestore.PageStore;
+import org.h2.pagestore.db.LobStorageBackend;
 import org.h2.result.Row;
 import org.h2.result.RowFactory;
 import org.h2.result.SimpleRow;
@@ -59,9 +61,7 @@ import org.h2.store.DataReader;
 import org.h2.store.FileLister;
 import org.h2.store.FileStore;
 import org.h2.store.FileStoreInputStream;
-import org.h2.store.LobStorageBackend;
 import org.h2.store.LobStorageFrontend;
-import org.h2.store.LobStorageMap;
 import org.h2.store.fs.FileUtils;
 import org.h2.util.IOUtils;
 import org.h2.util.IntArray;
