+++ h2/src/test/org/h2/test/unit/TestFile.java
@@ -7,9 +7,9 @@ package org.h2.test.unit;
 
 import java.util.Random;
 import org.h2.api.JavaObjectSerializer;
+import org.h2.pagestore.db.LobStorageBackend;
 import org.h2.store.DataHandler;
 import org.h2.store.FileStore;
-import org.h2.store.LobStorageBackend;
 import org.h2.store.fs.FileUtils;
 import org.h2.test.TestBase;
 import org.h2.util.SmallLRUCache;
